// src/app/models/ped_fat_model.ts
export default interface PedFatModel {
  PF_CODIGO: number;
  PF_DATA: string;
  PF_TABELA: string;
  PF_COD_PED: number;
  PF_CAMPO_FAT: string;
  PF_CAMPO_PED: string;
  PF_CLIENTE: string;
  PF_VALOR: number;
  PF_VALORPG: number;
  PF_COD_CLI: number;
  PF_FUN: number;
  PF_PARCELAS: number;
  PF_FAT: number;
  PF_VALORB: number;
  PF_DESCONTO: number;
  PF_DATAC: string;
  PF_CAMPO_DATAC: string;
  PF_TIPO: number;
  PF_FICHA?: number;
}